basic signup,login,logout

+++ src/App.js
@@ -3,28 +3,36 @@ import { BrowserRouter, Link, Switch } from "react-router-dom";
 import "./App.css";
 import AnonRoute from "./components/auth/AnonRoute";
 import PrivateRoute from "./components/auth/PrivateRoute";
-import { validateSession } from "./services/userService";
+import { validateSession, logout } from "./services/authService";
 import Home from "./views/Home";
 import Login from "./views/Login";
 import Signup from "./views/Signup";
 
 class App extends React.Component {
+  /** STATE: */
   state = {
     authenticated: false,
     user: {},
   };
+
+  /**
+   *  componentDidMount
+   */
   componentDidMount = () => {
     const accessToken = localStorage.getItem("accessToken");
     if (accessToken) {
       validateSession(accessToken)
         .then((response) => {
-          console.log(response, "RESPONSE");
           this.authenticate(response.session.userId);
         })
         .catch((err) => console.log(err));
     }
   };
 
+  /**
+   *
+   * @param {*} user
+   */
   authenticate = (user) => {
     this.setState({
       authenticated: true,
@@ -32,15 +40,31 @@ class App extends React.Component {
     });
   };
 
+  /**
+   *
+   */
   handleLogout = () => {
-    localStorage.clear();
-    this.setState({
-      authenticated: false,
-      user: {},
-    });
+    const token = localStorage.getItem("accessToken");
+    logout(token)
+      .then((resp) => {
+        console.log(" logout response: ", resp);
+        if (resp.success) {
+          localStorage.clear();
+          this.setState({
+            authenticated: false,
+            user: {},
+          });
+        }
+      })
+      .catch((err) => console.log(err));
   };
+
+  /**
+   *
+   */
   render() {
     const { authenticated } = this.state;
+
     return (
       <div className="App">
         <BrowserRouter>
